tidy up id newtypes with a macro, restructure crate
