+++ inselo/src/credential_provider.rs
@@ -1,5 +1,29 @@
-use crate::{error::Result, models::Credentials};
+use chrono::{DateTime, Utc};
+use secrecy::ExposeSecret;
+use secrecy::SecretString;
 
+use crate::{error::Result, models::ClientId};
+
+/// Implement this to source your credentials when they're needed. By default the token has a TTL
+/// of 1 hour.
 pub trait CredentialProvider {
     fn try_fetch_credentials(&self) -> Result<Credentials>;
+}
+
+#[derive(Debug, Clone)]
+pub struct Credentials {
+    pub client_id: ClientId,
+    pub client_secret: SecretString,
+}
+
+#[derive(Debug, Clone)]
+pub struct Token {
+    pub token: SecretString,
+    pub expires_at: DateTime<Utc>,
+}
+
+impl Token {
+    pub fn expose(&self) -> &str {
+        self.token.expose_secret()
+    }
 }
